Sped up tests of scan1() [mostly lm() is slow]
 - smaller number of phenotypes
 - more widely spaced pseudomarkers
 - for lm(), pick a smaller chromosome

The discrepancy between R/qtl and scan1(), for multiple phenotypes with
some missing data and use of weights, was due to a bug in R/qtl which is
now fixed. For now, revised the .travis.yml file to pull the devel
version of R/qtl from github. (Though I'm not totally sure if it works.)